requirements.txt
Django==5.0.7
mysqlclient
djangorestframework